requirements.txt
fastapi
uvicorn[standard]
httpx
openai>=1.30.1
python-dotenv
pyyaml
